HUB-217: Refactoring: Moved DegreeProgrammeCacheableMetadata to a separate file.
